Remove uncessary index.sh script
It seems that `index.sh` previously was being used in the testing
harness, as well as in the NPM package.json's `main` field. Some notes:

The `main` property should be reserved for JavaScript module IDs. This
doesn't make sense in the context of a shell CLI utility.

The `index.sh` script did the same things as `bin/runner`, but would
check to see if it shouldn't source `src/cli.sh`. So then there are two
cases:

1. `index.sh` sources `src/cli.sh`. Here, you may as well have just run
  `bin/runner` in the first place, skipping `index.sh`.
2. `index.sh` doesn't source `src/cli.sh`. Here, you gain nothing doing
  `source index.sh` over just `source src/runner.sh`, so again
  `index.sh` is unnecessary.

---

This file was used in conjunction with bash-require, which is not
supported now. (stylemistake)